% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce-as.data.frame-methods.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,IntegerRanges-method}
\alias{as.data.frame,Matrix-method}
\title{Coerce object to data.frame}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{IntegerRanges}(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{Matrix}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{row.names, optional}{Refer to \code{base::data.frame} for usage details.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame}.
}
\description{
Coerce object to data.frame
}
\note{
Updated 2022-02-08.
}
\examples{
data(
    IntegerRanges,
    sparseMatrix,
    package = "AcidTest"
)

## `IntegerRanges` to `data.frame` ====
from <- IntegerRanges
to <- as.data.frame(from)
head(to)

## `Matrix` to `data.frame` ====
from <- sparseMatrix
to <- as.data.frame(from)
head(to)
}
