% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce-as.data.table-methods.R
\name{as.data.table}
\alias{as.data.table}
\alias{as.data.table.DataFrame}
\alias{as.data.table.IntegerRanges}
\alias{as.data.table.GenomicRanges}
\title{Coerce object to a data.table}
\usage{
as.data.table(x, keep.rownames = FALSE, ...)

\method{as.data.table}{DataFrame}(x, keep.rownames = TRUE, ...)

\method{as.data.table}{IntegerRanges}(x, keep.rownames = TRUE, ...)

\method{as.data.table}{GenomicRanges}(x, keep.rownames = TRUE, ...)
}
\arguments{
\item{x}{Object.}

\item{keep.rownames}{Refer to \code{data.table::as.data.table} for usage details.}

\item{...}{Additional arguments.}
}
\value{
\code{data.table}.
}
\description{
Coerce object to a data.table
}
\details{
Our defined methods attempt to improve on the defaults in the data.table
package to ensure that row names are not dropped by default, which is a poor
default for bioinformatics. This is accomplished by setting
\code{keep.rownames = "rowname"} by default instead of \code{keep.rownames = NULL}.
Note that we're manually defining the \code{"rowname"} column instead of using
\code{TRUE}, to match the conventions used in our \code{as_tibble()} methods.

The package extends \code{as.data.table()} method support for these S4 classes:
\itemize{
\item \code{DataFrame} (from S4Vectors package).
\item \code{GenomicRanges} (from GenomicRanges package).
}
}
\note{
Updated 2022-02-08.
}
\examples{
data(
    DataFrame,
    GenomicRanges,
    IntegerRanges,
    package = "AcidTest"
)

## `DataFrame` to `data.table` ====
from <- DataFrame
to <- as.data.table(from)
print(to)

## `GenomicRanges` to `data.table` ====
from <- GenomicRanges
to <- as.data.table(from)
print(to)

## `IntegerRanges` to `data.table` ====
from <- IntegerRanges
to <- as.data.table(from)
print(to)
}
\seealso{
\itemize{
\item \code{data.table::as.data.table()}.
}
}
