% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coerce-as.DataFrame-methods.R
\name{as.DataFrame}
\alias{as.DataFrame}
\alias{as.DataFrame,GenomicRanges-method}
\alias{as.DataFrame,IntegerRanges-method}
\alias{as.DataFrame,Matrix-method}
\alias{as.DataFrame,SimpleList-method}
\alias{as.DataFrame,data.frame-method}
\alias{as.DataFrame,list-method}
\alias{as.DataFrame,matrix-method}
\title{Coerce object to S4 DataFrame}
\usage{
as.DataFrame(x, ...)

\S4method{as.DataFrame}{GenomicRanges}(x)

\S4method{as.DataFrame}{IntegerRanges}(x)

\S4method{as.DataFrame}{Matrix}(x)

\S4method{as.DataFrame}{SimpleList}(x, row.names = NULL)

\S4method{as.DataFrame}{data.frame}(x)

\S4method{as.DataFrame}{list}(x, row.names = NULL)

\S4method{as.DataFrame}{matrix}(x)
}
\arguments{
\item{x}{Object.}

\item{row.names}{Refer to \code{base::as.data.frame} for usage details.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Coerce object to S4 DataFrame
}
\note{
Updated 2022-05-02.
}
\examples{
data(
    data.table,
    sparseMatrix,
    tibble,
    package = "AcidTest"
)

## `Matrix` to `DataFrame` ====
from <- sparseMatrix
to <- as.DataFrame(from)
print(to)

## `data.table` to `DataFrame` ====
from <- data.table
to <- as.DataFrame(from)
head(to)

## `list` to `DataFrame` ====
## Use `as.DataFrame()` instead of `as()` for `list` class.
from <- list(
    "a" = list(c(1, 2), c(3, 4)),
    "b" = list(NULL, NULL)
)
to <- as.DataFrame(from)
print(to)

## `tbl_df` to `DataFrame` ====
from <- tibble
to <- as.DataFrame(from)
head(to)
}
