% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignAndSaveData.R
\name{assignAndSaveData}
\alias{assignAndSaveData}
\title{Assign and save data}
\usage{
assignAndSaveData(
  name,
  object,
  dir = getOption(x = "acid.save.dir", default = getwd()),
  ext = getOption(x = "acid.save.ext", default = "rds"),
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  compress = getOption(x = "acid.save.compress", default = TRUE),
  envir = parent.frame()
)
}
\arguments{
\item{name}{\code{character(1)}.
Desired variable name.}

\item{object}{Object.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{ext}{\code{character(1)}.
Output file format extension.

Supported arguments:
\itemize{
\item \code{"rds"}: R data serialized (RDS).
\item \code{"rda"}: R data (RDA).
}

RDS is preferred when saving single objects per file, which is always the
convention of \code{saveData()}, regardless of the extension used.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{compress}{\code{logical(1)} or \code{character(1)}.
These character strings are currently allowed for \code{\link[base:save]{save()}}:
\code{"gzip"}, \code{"bzip2"}, or \code{"xz"}.}

\item{envir}{\code{environment}.
Environment to use for assignment.
Defaults to \code{parent.frame()}, the calling environment.}
}
\value{
Invisible named \code{character(1)}.
File path.
}
\description{
Assigns a new object by name to the current working environment then saves
the newly assigned object, specified by the \code{"dir"} argument.
}
\note{
This function attempts to follow the same order as \code{assign()}.

Updated 2022-06-02.
}
\examples{
x <- 1L
dir <- AcidBase::tempdir2()
assignAndSaveData(
    name = "example",
    object = x,
    dir = dir,
    ext = "rds"
)
exists("example", inherits = FALSE)
file.exists(file.path(dir, "example.rds"))
AcidBase::unlink2(dir)
}
