% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/atomize-methods.R
\name{atomize}
\alias{atomize}
\alias{atomize,DataFrame-method}
\alias{atomize,Ranges-method}
\alias{atomize,data.frame-method}
\title{Atomize}
\usage{
atomize(object, ...)

\S4method{atomize}{DataFrame}(object)

\S4method{atomize}{Ranges}(object)

\S4method{atomize}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Non-atomic columns (e.g. \code{list} or complex S4 classes) will be dropped.
}
\description{
Ensure all columns in the object are \code{atomic}.
}
\details{
Keep only atomic columns. Complex columns won't write to disk as CSVs or work
with R Markdown functions, in some cases.
}
\note{
Updated 2021-10-12.
}
\examples{
df <- DataFrame(
    "a" = "a",
    "b" = I(list("a" = seq_len(3L)))
)
lapply(df, is.atomic)
x <- atomize(df)
}
