% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{brio-package}
\alias{brio}
\alias{brio-package}
\title{brio}
\description{
Extension of the rio package with improved support for biological data
structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steinbaugh/brio/}
  \item Report bugs at \url{https://github.com/steinbaugh/brio/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

}
