% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{brioCacheURL}
\alias{brioCacheURL}
\title{brio package cache URL}
\format{An object of class \code{character} of length 1.}
\usage{
brioCacheURL
}
\description{
brio package cache URL
}
\examples{
brioCacheURL
}
\keyword{internal}
