% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{brioTestsURL}
\alias{brioTestsURL}
\title{brio test data URL}
\format{An object of class \code{character} of length 1.}
\usage{
brioTestsURL
}
\description{
brio test data URL
}
\examples{
brioTestsURL
}
\keyword{internal}
