% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheUrl.R
\name{cacheUrl}
\alias{cacheUrl}
\title{Download and cache a file}
\usage{
cacheUrl(url, pkg = "pipette", update = FALSE, ask = FALSE, verbose = TRUE)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{pkg}{\code{character(1)}.
Package name.}

\item{update}{\code{logical(1)}.
Call \code{bfcneedsupdate} internally to see if URL needs an update.
Doesn't work reliably for all servers, so disabled by default.}

\item{ask}{\code{logical(1)}.
Ask if package cache needs to be created or if file needs to be updated.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}
}
\value{
\code{character(1)}.
Cached file path on disk.
}
\description{
Download and cache a file
}
\details{
Caching requires BiocFileCache and tools packages to be installed. If
BiocFileCache is not installed, URL will be downloaded as a temporary file
instead.
}
\note{
Updated 2023-09-26.
}
\examples{
url <- AcidBase::pasteUrl(pipetteTestsUrl, "biocfilecache-test.txt")
file <- cacheUrl(url)
print(file)
}
\seealso{
\itemize{
\item \code{BiocFileCache::bfcinfo()}.
}
}
