% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{compress}
\alias{compress}
\title{Compress or decompress a file}
\usage{
compress(
  file,
  ext = c("gz", "bz2", "xz", "zip"),
  remove = NULL,
  overwrite = NULL
)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{ext}{\code{character(1)}.
Compression file format extension. Uses \code{match.arg()} internally and
defaults to the first argument in the \code{character} vector.

Supported formats:
\itemize{
\item \code{gz}: gzip compression;
calls \code{gzfile()} internally.
\item \code{bz}: bzip2 (lzma) compression;
calls \code{bzfile()} internally.
\item \code{xz}: xz compression;
calls \code{xzfile()} internally.
\item \code{zip}: zip compression;
calls \code{zip()} or \code{unzip()} internally.
}}

\item{remove}{\code{logical(1)}.
Remove the input file once the output file is fully created and the
connection is closed.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}
}
\description{
Compress or decompress using gzip, bzip2, or xz compression.
}
\note{
For ZIP files, refer to \code{zip} and \code{unzip} in the utils package.

Updated 2021-09-24.
}
\examples{
## Create an example text file.
text <- c("hello", "world")
file <- "test.txt"
writeLines(text = text, con = file)
readLines(con = file)

## Apply gzip compression.
gzfile <- compress(file = file, ext = "gz", remove = TRUE, overwrite = TRUE)
gzfile
readLines(con = gzfile)
## When `remove = TRUE`, the original input file will be removed.
file.exists(file)

## Decompress the gzipped file.
file <- decompress(file = gzfile, remove = TRUE, overwrite = TRUE)
file

unlink(file)
}
\seealso{
\itemize{
\item \code{help("connections")}
\item \code{R.utils::compressFile}, \code{R.utils::decompressFile}.
\item \code{utils::zip}, \code{utils::unzip}.
\item \code{utils::tar}, \code{utils::untar}.
\item \code{memCompress}, \code{memDecompress}.
}
}
