% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{compressExtPattern}
\alias{compressExtPattern}
\title{Compression extension pattern}
\format{An object of class \code{character} of length 1.}
\usage{
compressExtPattern
}
\description{
Compression extension pattern
}
\examples{
compressExtPattern
}
\keyword{datasets}
