% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/decode-methods.R
\name{decode}
\alias{decode}
\alias{decode,DFrame-method}
\alias{decode,Ranges-method}
\title{Decode data that uses run-length encoding}
\usage{
decode(x, ...)

\S4method{decode}{DFrame}(x, j = NULL)

\S4method{decode}{Ranges}(x)
}
\arguments{
\item{x}{Object.}

\item{j}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:

\if{html}{\out{<div class="sourceCode">}}\preformatted{help(topic = "Extract", package = "base")
}\if{html}{\out{</div>}}}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Columns will be decoded and no longer \code{Rle} class.
}
\description{
Decode data that uses run-length encoding
}
\note{
Updated 2023-09-20.
}
\examples{
data(DFrame, package = "AcidTest")

## DFrame ====
object <- DFrame
object <- encode(object)
lapply(object, class)
object <- decode(object)
lapply(object, class)
}
\seealso{
\itemize{
\item \code{S4Vectors::decode}.
\item \code{S4Vectors::Rle}.
}
}
