% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{sanitizeColData}
\alias{sanitizeRowData}
\alias{sanitizeRowRanges}
\title{Deprecated functions}
\usage{
sanitizeColData(...)

sanitizeRowData(...)

sanitizeRowRanges(...)
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\keyword{internal}
