% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download a file from the Internet}
\usage{
download(url, destfile, quiet = FALSE, ...)
}
\arguments{
\item{url}{\code{character(1)}.
URL.}

\item{destfile}{\code{character(1)}.
Destination file.}

\item{...}{Passthrough arguments to \code{\link[utils:download.file]{utils::download.file()}}.}
}
\value{
\code{character(1)}.
Destination file path.
Note that this differs from \code{download.file}, which returns a status code
(e.g. \code{0} for success) instead.
}
\description{
Download a file from the Internet
}
\details{
Unlike \code{\link[utils:download.file]{utils::download.file()}}, intentionally errors on any download
failure. Automatically sets a long time out internally via
\code{options(timeout = 9999L)}, to avoid download issues related to \code{timeout}
configuration in \code{Rprofile}.
}
\note{
Updated 2020-12-10.
}
\examples{
url <- "https://bioconductor.org/bioc-version"
destfile <- "bioc-version.txt"
out <- download(url = url, destfile = destfile)
print(out)
file.remove(out)
}
