% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/droplevels-methods.R
\name{droplevels}
\alias{droplevels}
\alias{droplevels,DFrame-method}
\alias{droplevels,Ranges-method}
\title{Drop unused levels from factors}
\usage{
droplevels(x, ...)

\S4method{droplevels}{DFrame}(x)

\S4method{droplevels}{Ranges}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
The function \code{droplevels} is used to drop unused levels from a
  \code{\link[base]{factor}} or, more commonly, from factors in a data frame.
}
\note{
Updated 2021-02-03.
}
\examples{
data(GRanges, package = "AcidTest")

## Ranges ====
object <- GRanges
object <- droplevels(object)
print(object)
}
