% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/droplevels2-methods.R
\name{droplevels2}
\alias{droplevels2}
\alias{droplevels2,DFrame-method}
\alias{droplevels2,Ranges-method}
\alias{droplevels2,Rle-method}
\alias{droplevels2,data.frame-method}
\alias{droplevels2,factor-method}
\title{Drop factor levels}
\usage{
droplevels2(x, ...)

\S4method{droplevels2}{DFrame}(x)

\S4method{droplevels2}{Ranges}(x)

\S4method{droplevels2}{Rle}(x)

\S4method{droplevels2}{data.frame}(x)

\S4method{droplevels2}{factor}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Drop factor levels
}
\note{
Updated 2023-09-12.
}
\examples{
data(GRanges, package = "AcidTest")

## Ranges ====
object <- GRanges
object <- droplevels2(object)
print(object)

## Rle ====
object <- factor(c("a", "a", "b", "b"), levels = c("a", "b", "c"))
object <- S4Vectors::Rle(object)
print(object)
object <- droplevels2(object)
print(object)

## factor ====
object <- factor(c("a", "a", "b", "b"), levels = c("a", "b", "c"))
print(object)
object <- droplevels2(object)
print(object)
}
