% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode-methods.R
\name{encode}
\alias{encode}
\alias{encode,DataFrame-method}
\alias{encode,Ranges-method}
\alias{encode,SummarizedExperiment-method}
\title{Encode using run-length encoding}
\usage{
encode(x, ...)

\S4method{encode}{DataFrame}(x)

\S4method{encode}{Ranges}(x)

\S4method{encode}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
All \code{atomic} columns will be encoded to \code{Rle} S4 class.
}
\description{
Encode using run-length encoding
}
\note{
Updated 2019-07-20.
}
\examples{
## DataFrame ====
binary <- seq(from = 0L, to = 1L)
df <- S4Vectors::DataFrame(
    a = rep(x = binary, times = 50L),
    b = rep(x = binary, each = 50L)
)
lapply(df, class)
x <- encode(df)
lapply(x, class)
}
\seealso{
\code{\link[S4Vectors:Rle]{S4Vectors::Rle()}}.
}
