% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/factorize-methods.R
\name{factorize}
\alias{factorize}
\alias{factorize,DataFrame-method}
\alias{factorize,data.frame-method}
\title{Factorize}
\usage{
factorize(object, ...)

\S4method{factorize}{DataFrame}(object)

\S4method{factorize}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
All columns will be coerced to \code{factor}.
}
\description{
Coerce all columns in the object to \code{factor} (i.e. \code{stringsAsFactors}).
}
\note{
Updated 2022-08-17.
}
\examples{
## DataFrame ====
df <- S4Vectors::DataFrame(
    "a" = letters[seq_len(5L)],
    "b" = seq_len(5L)
)
x <- factorize(df)
}
\seealso{
\itemize{
\item Legacy \code{stringsAsFactors} approach for \code{data.frame} import.
}
}
