% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\alias{fileExists}
\alias{fileCreate}
\alias{fileRemove}
\alias{fileAppend}
\alias{fileCopy}
\alias{fileSymlink}
\title{File manipulation}
\usage{
fileExists(...)

fileCreate(...)

fileRemove(...)

fileAppend(from, to)

fileCopy(from, to, overwrite = FALSE)

fileSymlink(from, to, overwrite = FALSE)
}
\description{
File manipulation
}
\note{
Updated 2019-10-17.
}
\examples{
unlink(c("a.txt", "b.txt", "c.txt"))
fileCreate("a.txt")
fileCopy(from = "a.txt", to = "b.txt")
fileAppend(from = "a.txt", to = "b.txt")
## Symbolic links don't work on Windows.
if (.Platform$OS.type != "windows") {
    fileSymlink(from = "a.txt", to = "c.txt")
}
fileRemove("b.txt")
fileExists("a.txt", "b.txt", "c.txt")
unlink(c("a.txt", "b.txt", "c.txt"))
}
\seealso{
\code{help("files")}
\itemize{
\item \code{file.exists()}.
\item \code{file.create()}.
\item \code{file.remove()}.
\item \code{file.rename()}.
\item \code{file.append()}.
\item \code{file.copy()}.
\item \code{file.symlink()}.
\item \code{file.link()}.
\item \code{goalie::isFile()}.
\item \code{goalie::allAreFiles()}.
}
}
