% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillLines.R
\name{fillLines}
\alias{fillLines}
\title{Fill delimited lines with NA values}
\usage{
fillLines(object, format = c("csv", "tsv"))
}
\arguments{
\item{object}{\code{character}.}

\item{format}{\code{character(1)}.
Delimited input format.}
}
\value{
\code{character}.
Modified character vector, with consistent number of delimited values.
}
\description{
Fill delimited lines with NA values
}
\note{
Updated 2023-09-27.
}
\examples{
## character ====
## CSV format.
object <- c(
    "\"aaa\",\"bbb\",\"ccc\",\"ddd\"",
    "\"eee\",\"fff\",\"ggg\"",
    "\"hhh\",\"iii\""
)
print(object)
object <- fillLines(object, format = "csv")
print(object)
con <- textConnection(object)
object <- import(con, format = "csv")
print(object)
close(con)

## TSV format.
object <- c(
    "aaa\tbbb\tccc\tddd",
    "eee\tfff\tggg",
    "hhh\tiii"
)
print(object)
object <- fillLines(object, format = "tsv")
print(object)
con <- textConnection(object)
object <- import(con, format = "tsv")
print(object)
close(con)
}
\seealso{
\itemize{
\item \code{textConnection}: For import approach after modification.
}
}
