% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURL.R
\name{getURL}
\alias{getURL}
\title{Download and import a URL}
\usage{
getURL(url, quiet = getOption(x = "acid.quiet", default = FALSE))
}
\value{
\code{character}.
Source code lines.
}
\description{
Download and import a URL
}
\note{
Updated 2023-09-19.
}
\examples{
url <- "https://ftp.ncbi.nlm.nih.gov/genomes/"
if (goalie::isAnExistingURL(url)) {
    lines <- getURL(url)
    print(head(lines, n = 7L))
}
}
