% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURLDirList.R
\name{getURLDirList}
\alias{getURLDirList}
\title{Get remote URL directory listing}
\usage{
getURLDirList(url, pattern = NULL, absolute = FALSE)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{pattern}{\code{character(1)}.
Regular expression pattern to use for matching.
Passes to \code{grepl} function internally.}

\item{absolute}{\code{logical(1)}.
Return absolute path.}
}
\value{
\code{character}.
Simple directory contents return, including both files and subdirectories.
}
\description{
Get remote URL directory listing
}
\details{
Currently, only FTP servers are supported.
}
\note{
Updated 2023-09-17.
}
\examples{
url <- "ftp://ftp.ncbi.nlm.nih.gov/genomes/"
if (goalie::isAnExistingURL(url)) {
    x <- getURLDirList(url)
    print(x)
}
}
\seealso{
\itemize{
\item \code{curlGetHeaders}.
}
}
