% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRemoteData.R
\name{loadRemoteData}
\alias{loadRemoteData}
\title{Load remote data}
\usage{
loadRemoteData(url, envir = globalenv())
}
\arguments{
\item{url}{\code{character}.
Remote URL file path(s) to R data.}

\item{envir}{\code{environment}.
Environment.}
}
\value{
Invisible named \code{character}.
Local object name as the name, and the remote URL as the value.
}
\description{
Load a remote R binary file. This function is vectorized and supports
multiple URLs in a single call.
}
\examples{
url <- paste(brioTestsURL, "rnaseq_counts.rds", sep = "/")
print(url)
x <- loadRemoteData(url)
print(x)
}
