% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/matchRownameColumn-methods.R
\name{matchRownameColumn}
\alias{matchRownameColumn}
\alias{matchRownameColumn,data.frame-method}
\alias{matchRownameColumn,DataFrame-method}
\title{Match row name column}
\usage{
matchRownameColumn(object, ...)

\S4method{matchRownameColumn}{data.frame}(
  object,
  choices = c("rn", "row.name", "row.names", "rowname", "rownames")
)

\S4method{matchRownameColumn}{DataFrame}(
  object,
  choices = c("rn", "row.name", "row.names", "rowname", "rownames")
)
}
\arguments{
\item{object}{Object.}

\item{choices}{\code{character}.
Column name choices to use internally for matching.
Note that case-insensitive matching is performed against \code{make.names()}
return internally. Either dots (".") or underscores ("_") used as word
separators will match.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)} or \code{NULL}.
\itemize{
\item data.table: \code{"rn"}.
\item tibble: \code{"rowname"}.
}
}
\description{
Automatically detect row names column, if defined.
}
\details{
The data.table package uses "rn" by default, whereas tibble uses "rowname".
}
\note{
Updated 2020-01-18.
}
\examples{
data(data.table, tbl_df, package = "AcidTest")

## data.table ====
matchRownameColumn(data.table)

## tbl_df ====
matchRownameColumn(tbl_df)
}
