% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metadata2-methods.R
\name{metadata2}
\alias{metadata2}
\alias{metadata2<-}
\alias{metadata2,ANY,character-method}
\alias{metadata2,Annotated,character-method}
\alias{metadata2<-,ANY,character-method}
\alias{metadata2<-,Annotated,character-method}
\title{Metadata}
\usage{
metadata2(x, which, ...)

metadata2(x, which, ...) <- value

\S4method{metadata2}{ANY,character}(x, which)

\S4method{metadata2}{Annotated,character}(x, which)

\S4method{metadata2}{ANY,character}(x, which) <- value

\S4method{metadata2}{Annotated,character}(x, which) <- value
}
\arguments{
\item{x}{Object.}

\item{which}{\code{character(1)}.
A non-empty character string specifying which attribute is to be accessed.
Note that positional \code{numeric} arguments are currently not allowed.}

\item{value}{\code{ANY}.
Metadata values to assign into slot, defined by \code{which}.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{metadata2()}: Metadata. Returns \code{NULL} on \code{which} match failure, similar
to \code{\link[base:attr]{attr()}} and \code{\link[S4Vectors:Annotated-class]{metadata()}}.
\item \verb{metadata2<-()}: Modified object.
}
}
\description{
Dynamically handles metadata assignment and extraction in a similar fashion
for both S3 and S4 objects.
}
\note{
Updated 2020-01-18.
}
\section{S3 class}{


Internally slots into \code{\link[base:attributes]{attributes()}}.
Attribute is also accessible via \code{attr(object, which)}.
}

\section{S4 class}{


Requires that object extends \code{\link[S4Vectors:Annotated-class]{Annotated}} class,
which supports \code{\link[S4Vectors:Annotated-class]{metadata()}}.

Internally slots into \code{\link[S4Vectors:Annotated-class]{metadata()}}.
Attribute is also accessible via \code{metadata(object)[[which]]}.
}

\examples{
## S3 ====
x <- data.frame()
metadata2(x, which = "A") <- "B"
metadata2(x, which = "A")
attr(x, which = "A")

## S4 (extending Annotated) ====
x <- S4Vectors::DataFrame()
metadata2(x, which = "A") <- "B"
metadata2(x, which = "A")
S4Vectors::metadata(x)[["A"]]
}
