% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{dir}{\code{character(1)}.
Directory path.}

\item{envir}{\code{environment}.
Environment.}

\item{file}{\code{character(1)}.
File path.}

\item{object}{Object.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{x}{Object.}
}
\description{
Parameters
}
\keyword{internal}
