% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pipette-package}
\alias{pipette}
\alias{pipette-package}
\title{pipette}
\description{
Input/output functions for biological data formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pipette.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/pipette/}
  \item Report bugs at \url{https://github.com/acidgenomics/pipette/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
