% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PipetteFile-class}
\alias{PipetteFile-class}
\alias{DelimFile-class}
\alias{ExcelFile-class}
\alias{LinesFile-class}
\alias{RioHandoffFile-class}
\alias{RtracklayerHandoffFile-class}
\alias{BcbioCountsFile-class}
\alias{CSVFile-class}
\alias{FASTAFile-class}
\alias{FASTQFile-class}
\alias{GMTFile-class}
\alias{GMXFile-class}
\alias{GRPFile-class}
\alias{JSONFile-class}
\alias{MTXFile-class}
\alias{OBOFile-class}
\alias{PZFXFile-class}
\alias{RDataFile-class}
\alias{RDSFile-class}
\alias{TableFile-class}
\alias{TSVFile-class}
\alias{YAMLFile-class}
\title{File extension classes}
\description{
Currently intended for use with \code{import} function.
}
\details{
Extends \code{BiocFile} defined in BiocIO package.
}
\note{
Updated 2022-08-19.
}
\section{Primary classes}{

\itemize{
\item \code{BcbioCountsFile}: bcbio-nextgen counts file.
\item \code{CSVFile}: Comma-separated values file (CSV).
\item \code{FASTAFile}: FASTA file.
\item \code{FASTQFile}: FASTQ file.
\item \code{GMTFile}: Gene matrix transposed file (GMT).
\item \code{GMXFile}: Gene matrix file (GMX).
\item \code{GRPFile}: Gene set file (GRP).
\item \code{JSONFile}: JSON file.
\item \code{MTXFile}: MatrixMarket exchange file (MTX).
\item \code{OBOFile}: Open Biomedical Ontologies file (OBO).
\item \code{PZFXFile}: GraphPad Prism file (PZFX).
\item \code{RDSFile}: R data file containing a single, serialized object (RDS).
\item \code{RDataFile}: R Data file containing multiple objects (RData/RDA).
\item \code{TSVFile}: Tab-separated values file (TSV).
\item \code{TableFile}: Base R table file (TXT).
\item \code{YAMLFile}: YAML file.
}
}

\section{\code{DelimFile-class}}{


Delimited file.

File extension group supporting:
\itemize{
\item CSV: Comma-separated values
\item TSV: Tab-separated values
\item Base R table (TXT)
}
}

\section{\code{ExcelFile-class}}{


Microsoft Excel file.

File extension group supporting:
\itemize{
\item XLS
\item XLSB
\item XLSX
}
}

\section{\code{LinesFile-class}}{


Source code lines file.

File extension group supporting:
\itemize{
\item BASH: Bash shell script
\item LOG: Log file
\item MD: Markdown file
\item PY: Python script
\item R: R script
\item RMD: R Markdown file
\item SH: Shell script (e.g. Bash, POSIX)
\item ZSH: Zsh shell script
}
}

\section{\code{RioHandoffFile-class}}{


File extension supported by \code{rio::import}.

File extension group supporting:
\itemize{
\item ARFF: Weka Attribute-Relation File Format
\item DBF: dBase Database File
\item DIF: Data Interchange Format
\item DTA: Stata
\item FWF: Fixed-width formatting
\item MAT: Matlab
\item MTP: Minitab
\item ODS: OpenDocument (LibreOffice)
\item POR: SPSS
\item PSV: Pipe-separated values
\item SAS7BDAT: SASS
\item SAV: SPSS
\item SYD: Systat
\item REC: Epi Info
\item XPT: SASS
}
}

\section{\code{RtracklayerHandoffFile-class}}{


File extension supported by \code{rtracklayer::import()}.

File extension group supporting:
\itemize{
\item BED, BED15, BEDGRAPH, BEDPE
\item BIGWIG, BW, WIG
\item GFF, GFF1, GFF2, GFF3, GTF
\item BROADPEAK, NARROWPEAK
}
}

\seealso{
\itemize{
\item \code{getGeneric}.
\item \code{BiocIO::FileForFormat}.
}
}
