% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{data.table}
\alias{getURL}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{data.table}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}
}}

