% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/removeNa-methods.R
\name{removeNa}
\alias{removeNa}
\alias{removeNa,DFrame-method}
\alias{removeNa,Matrix-method}
\alias{removeNa,atomic-method}
\alias{removeNa,data.frame-method}
\alias{removeNa,matrix-method}
\title{Remove rows and columns containing only NA values}
\usage{
removeNa(object, ...)

\S4method{removeNa}{DFrame}(object)

\S4method{removeNa}{Matrix}(object)

\S4method{removeNa}{atomic}(object)

\S4method{removeNa}{data.frame}(object)

\S4method{removeNa}{matrix}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Sanitized object.
}
\description{
Remove rows and columns containing only NA values
}
\note{
Updated 2021-10-14.
}
\examples{
## atomic ====
removeNa(c("hello", "world", NA))
removeNa(c(1, 2, NA))

## matrix ====
from <- matrix(
    data = c(1, NA, 3, NA, NA, NA, 2, NA, 4),
    nrow = 3,
    ncol = 3
)
print(from)
to <- removeNa(from)
print(to)

## DFrame ====
from <- S4Vectors::DataFrame(
    "a" = c("A", NA, "C"),
    "b" = c(NA, NA, NA),
    "c" = c("B", NA, "D")
)
print(from)
to <- removeNa(from)
print(to)
}
