% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeColData.R
\name{sanitizeColData}
\alias{sanitizeColData}
\title{Sanitize column data}
\usage{
sanitizeColData(object)
}
\arguments{
\item{object}{\code{DataFrame}.}
}
\value{
\code{DataFrame}.
Contains only \code{atomic} columns.
}
\description{
Standardize data describing the colummns (e.g. samples or cells), keeping
only \code{atomic} columns.
}
\details{
Always applied:
\itemize{
\item Non-\code{atomic} columns will be flattened to \code{atomic} if possible, and
otherwise dropped.
}
}
\section{Sample data}{


The following conventions are enforced:
\itemize{
\item Required columns: \code{sampleName}.
\item Blacklisted columns: \code{interestingGroups}, \code{sampleID}.
\item All columns get coerced to \code{factor} and \code{\link[base:droplevels]{droplevels()}}
is applied.
}
}

