% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sanitizeNa-methods.R
\name{sanitizeNa}
\alias{sanitizeNa}
\alias{sanitizeNa,DFrame-method}
\alias{sanitizeNa,atomic-method}
\alias{sanitizeNa,character-method}
\alias{sanitizeNa,data.frame-method}
\alias{sanitizeNa,factor-method}
\title{Sanitize NA values}
\usage{
sanitizeNa(object, ...)

\S4method{sanitizeNa}{DFrame}(object)

\S4method{sanitizeNa}{atomic}(object)

\S4method{sanitizeNa}{character}(object)

\S4method{sanitizeNa}{data.frame}(object)

\S4method{sanitizeNa}{factor}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Sanitized to contain proper \code{NA} values.
}
\description{
Standardize empty strings (\code{""}), character NAs (\code{"NA"}), and \code{NULL} values
inside a character vector to \code{NA_character_}. Other \code{atomic} data types are
returned unmodified.
}
\note{
Updated 2023-09-20.
}
\examples{
## character ====
from <- as.character(c(1L, "x", "", "NA", "NULL"))
print(from)
to <- sanitizeNa(from)
print(to)

## DFrame ====
from <- S4Vectors::DataFrame(
    "a" = c("foo", ""),
    "b" = c(NA, "bar"),
    row.names = c("c", "d")
)
print(from)
to <- sanitizeNa(from)
print(to)
}
