% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Save data}
\usage{
saveData(..., dir = getOption("acid.save.dir", default = "."),
  ext = getOption("acid.save.ext", default = c("rds", "rda")),
  overwrite = getOption("acid.overwrite", default = TRUE),
  compress = getOption("acid.save.compress", default = TRUE),
  list = NULL, envir = parent.frame())
}
\arguments{
\item{...}{Object names.
Note that these arguments are interpreted as symbols using non-standard
evaluation for convenience during interactive use, and \emph{must not be
quoted}.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{ext}{\code{character(1)}.
R data serialized (\code{RDS}) or R data (\code{RDA}, \code{RDATA}). RDS is
preferred when saving single objects per file, which is always the
convention of \code{\link[=saveData]{saveData()}}, regardless of the extension used.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{compress}{logical or character string specifying whether saving
    to a named file is to use compression.  \code{TRUE} corresponds to
    \command{gzip} compression, and character strings \code{"gzip"},
    \code{"bzip2"} or \code{"xz"} specify the type of
    compression.  Ignored when \code{file} is a connection and
    for workspace format version 1.}

\item{list}{\code{character}.
A character vector containing the names of objects to be saved.
Note that this approach differs from \code{\link[base:save]{save()}} in that the
objects are saved individually to disk, instead of inside a single R data
file. Requires objects to be defined in \code{\link[base:globalenv]{globalenv()}}.}

\item{envir}{\code{environment}.
Environment.}
}
\value{
Invisible named \code{character}.
File paths.
}
\description{
Wrapper for \code{\link[base:save]{save()}} supporting quick, interactive saving of
object names passed as symbols.
}
\details{
This function always saves each object into a separate file rather than
combining multiple objects into a single file.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation. It will \strong{overwrite} existing files
on disk, following the same conventions as \code{\link[base:save]{save()}}.
}
\examples{
dir <- "example"

## Interactive mode ====
## Note that this method uses non-standard evaluation.
a <- 1
b <- 2
saveData(a, b, dir = dir)
sort(list.files(dir))

## Clean up.
unlink(dir, recursive = TRUE)

## List mode ====
## Note that this method uses standard evaluation.
## Use this approach inside of functions.
a <- 1
b <- 2
list <- c("a", "b")
saveData(list = list, dir = dir)
sort(list.files(dir))

## Clean up.
unlink(dir, recursive = TRUE)
}
\seealso{
\itemize{
\item \code{\link[base:save]{save()}}
\item \code{\link[base:saveRDS]{saveRDS()}}.
}
}
