% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{transmit}
\alias{transmit}
\title{Transmit files from a remote FTP server}
\usage{
transmit(
  remoteDir,
  localDir = getwd(),
  pattern = NULL,
  rename = NULL,
  compress = FALSE,
  download = TRUE
)
}
\arguments{
\item{remoteDir}{\code{character(1)}.
Remote FTP directory path.}

\item{localDir}{\code{character(1)}.
Directory where to save files locally.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{rename}{\code{character(1)} or \code{NULL}.
Rename the local files (including suffix), if desired.}

\item{compress}{\code{logical(1)}.
Compress the downloaded files.}

\item{download}{\code{logical(1)}.
Download files (default) or only return the matching URL(s).}
}
\value{
Invisible \code{character}.
Local file paths.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\details{
Requires RCurl package to be installed.
}
\note{
Updated 2023-09-14.
}
\examples{
try({
    localDir <- AcidBase::tempdir2()
    readme <- transmit(
        remoteDir = "ftp://ftp.ncbi.nlm.nih.gov/genomes/",
        localDir = localDir,
        pattern = "^README\\\\.txt$",
        rename = "ncbi-readme.txt",
        compress = FALSE
    )
    basename(readme)
    file.exists(readme)
    AcidBase::unlink2(localDir)
})
}
