% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/unfactorize-methods.R
\name{unfactorize}
\alias{unfactorize}
\alias{unfactorize,factor-method}
\alias{unfactorize,data.frame-method}
\alias{unfactorize,DFrame-method}
\title{Unfactorize}
\usage{
unfactorize(object, ...)

\S4method{unfactorize}{factor}(object)

\S4method{unfactorize}{data.frame}(object)

\S4method{unfactorize}{DFrame}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object, with \code{factor} coerced back to atomic vector.
}
\description{
Coerce all columns in the object from \code{factor}.
}
\note{
Updated 2023-09-20.
}
\examples{
## factor ====
object <- as.factor(c(100L, 100L, 101L, 101L))
## Coercion with `as.numeric` or `as.integer` don't work the way we want.
print(as.numeric(object))
print(as.integer(object))
}
\seealso{
\itemize{
\item Warning section of \code{help("factor")}.
\item R FAQ 7.10: How do I convert factors to numeric?
\item https://stackoverflow.com/questions/3418128/
\item forcats package.
}
}
