% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/as.DataFrame-methods.R
\name{as.DataFrame}
\alias{as.DataFrame}
\alias{as.DataFrame,GRanges-method}
\alias{as.DataFrame,IRanges-method}
\alias{as.DataFrame,Matrix-method}
\alias{as.DataFrame,SimpleList-method}
\alias{as.DataFrame,data.frame-method}
\alias{as.DataFrame,list-method}
\alias{as.DataFrame,matrix-method}
\title{Coerce object to S4 DataFrame}
\usage{
as.DataFrame(x, ...)

\S4method{as.DataFrame}{GRanges}(x)

\S4method{as.DataFrame}{IRanges}(x)

\S4method{as.DataFrame}{Matrix}(x)

\S4method{as.DataFrame}{SimpleList}(x)

\S4method{as.DataFrame}{data.frame}(x)

\S4method{as.DataFrame}{list}(x)

\S4method{as.DataFrame}{matrix}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Coerce object to S4 DataFrame
}
\note{
Updated 2023-02-22.
}
\examples{
data(GRanges, sparseMatrix, package = "AcidTest")

## `GRanges` to `DFrame` ====
from <- GRanges
to <- as.DataFrame(from)
print(to)

## `Matrix` to `DFrame` ====
from <- sparseMatrix
to <- as.DataFrame(from)
print(to)

## `list` to `DFrame` ====
from <- list(
    "a" = list(c(1, 2), c(3, 4)),
    "b" = list(NULL, NULL)
)
to <- as.DataFrame(from)
print(to)
}
