% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignAndSaveData.R
\name{assignAndSaveData}
\alias{assignAndSaveData}
\title{Assign and save data}
\usage{
assignAndSaveData(name, object, envir = parent.frame(),
  dir = getOption("acid.save.dir", default = "."),
  ext = getOption("acid.save.ext", default = c("rds", "rda")),
  overwrite = getOption("acid.overwrite", default = TRUE),
  compress = getOption("acid.save.compress", default = TRUE))
}
\arguments{
\item{name}{\code{character(1)}.
Desired variable name.}

\item{object}{Object.}

\item{envir}{\code{environment}.
Environment to use for assignment.
Defaults to \code{\link[=parent.frame]{parent.frame()}}, the calling environment.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{ext}{\code{character(1)}.
R data serialized (\code{RDS}) or R data (\code{RDA}, \code{RDATA}). RDS is
preferred when saving single objects per file, which is always the
convention of \code{\link[=saveData]{saveData()}}, regardless of the extension used.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{compress}{\code{logical(1)} or \code{character(1)}.
These character strings are currently allowed for \code{\link[base:save]{save()}}:
\code{"gzip"}, \code{"bzip2"}, or \code{"xz"}.}
}
\value{
Invisible named \code{character(1)}.
File path.
}
\description{
Assigns a new object by name to the current working environment then saves
the newly assigned object, specified by the \code{"dir"} argument.
}
\note{
This function attempts to follow the same order as
\code{\link[base:assign]{assign()}}.

Updated 2019-09-06.
}
\examples{
x <- 1L
assignAndSaveData(
    name = "example",
    object = x,
    dir = ".",
    ext = "rds"
)
exists("example", inherits = FALSE)
file.exists("example.rds")

## Clean up.
rm(example)
unlink("example.rds")
}
