% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{brio-package}
\alias{brio}
\alias{brio-package}
\title{brio}
\description{
Extension of the rio and rtracklayer packages with improved support for
biological data structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brio.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/brio/}
  \item Report bugs at \url{https://github.com/acidgenomics/brio/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
