% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheURL.R
\name{cacheURL}
\alias{cacheURL}
\title{Download and cache a file}
\usage{
cacheURL(
  url,
  pkg = "BiocFileCache",
  update = FALSE,
  ask = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{pkg}{\code{character(1)}.
Package name.}

\item{update}{\code{logical(1)}.
Call \code{bfcneedsupdate} internally to see if URL needs an update.
Doesn't work reliably for all servers, so disabled by default.}

\item{ask}{\code{logical(1)}.
Ask if package cache needs to be created or if file needs to be updated.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}
}
\value{
\code{character(1)}.
Cached file path on disk.
}
\description{
Download and cache a file
}
\details{
Requires BiocFileCache and tools packages to be installed.
}
\note{
Updated 2023-09-14.
}
\examples{
url <- AcidBase::pasteURL(
    pipetteTestsURL,
    "biocfilecache-test.txt",
    protocol = "none"
)
file <- cacheURL(url)
print(file)
}
\seealso{
\itemize{
\item \code{BiocFileCache::bfcinfo()}.
}
}
