% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/decode-methods.R
\name{decode}
\alias{decode}
\alias{decode,DataFrame-method}
\alias{decode,Ranges-method}
\title{Decode data that uses run-length encoding}
\usage{
decode(x, ...)

\S4method{decode}{DataFrame}(x)

\S4method{decode}{Ranges}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Columns will be decoded and no longer \code{Rle} class.
}
\description{
Decode data that uses run-length encoding
}
\note{
Updated 2022-02-04.
}
\examples{
data(DataFrame, package = "AcidTest")

## DataFrame ====
object <- DataFrame
object <- encode(object)
lapply(object, class)
object <- decode(object)
lapply(object, class)
}
\seealso{
\itemize{
\item \code{S4Vectors::decode()}.
}
}
