% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/decode-methods.R
\name{decode}
\alias{decode}
\alias{decode,DataFrame-method}
\alias{decode,Ranges-method}
\title{Decode column data that uses run-length encoding}
\usage{
decode(x, ...)

\S4method{decode}{DataFrame}(x)

\S4method{decode}{Ranges}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Columns will be decoded and no longer \code{Rle} class.
}
\description{
Decode (run-length encoding)
}
\note{
Updated 2021-06-09.
}
\examples{
data(DFrame, package = "AcidTest")
df <- DFrame

## DataFrame ====
df <- encode(df)
lapply(df, class)
x <- decode(df)
lapply(x, class)
}
\seealso{
\code{S4Vectors::decode()}.
}
