% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{sanitizeColData}
\alias{sanitizeRowData}
\alias{sanitizeRowRanges}
\title{Defunct functions}
\usage{
sanitizeColData(...)

sanitizeRowData(...)

sanitizeRowRanges(...)
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\keyword{internal}
