% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels-methods.R
\name{droplevels}
\alias{droplevels}
\alias{droplevels,Ranges-method}
\alias{droplevels,SummarizedExperiment-method}
\title{Drop unused levels from factors}
\usage{
droplevels(x, ...)

\S4method{droplevels}{Ranges}(x)

\S4method{droplevels}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
The function \code{droplevels} is used to drop unused levels from a
  \code{\link[base]{factor}} or, more commonly, from factors in a data frame.
}
\note{
Updated 2020-04-10.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")

## Ranges ====
rse <- RangedSummarizedExperiment
object <- SummarizedExperiment::rowRanges(rse)
S4Vectors::droplevels(object)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
S4Vectors::droplevels(object)
}
