% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/droplevels2-methods.R
\name{droplevels2}
\alias{droplevels2}
\alias{droplevels2,DataFrame-method}
\alias{droplevels2,Ranges-method}
\title{Drop factor levels}
\usage{
droplevels2(x, ...)

\S4method{droplevels2}{DataFrame}(x)

\S4method{droplevels2}{Ranges}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Drop factor levels
}
\note{
Updated 2023-02-22.
}
\examples{
data(GenomicRanges, package = "AcidTest")

## Ranges ====
object <- GenomicRanges
object <- droplevels2(object)
print(object)
}
