% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/encode-methods.R
\name{encode}
\alias{encode}
\alias{encode,DFrame-method}
\alias{encode,Ranges-method}
\title{Encode using run-length encoding}
\usage{
encode(x, ...)

\S4method{encode}{DFrame}(x, j = NULL)

\S4method{encode}{Ranges}(x)
}
\arguments{
\item{x}{Object.}

\item{j}{\code{vector}.
Column names or positions to evaluate.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
All \code{atomic} columns will be encoded to \code{Rle} S4 class.
}
\description{
Encode using run-length encoding
}
\note{
Updated 2023-09-20.
}
\examples{
## DFrame ====
binary <- seq(from = 0L, to = 1L)
object <- S4Vectors::DataFrame(
    "a" = rep(x = binary, times = 50L),
    "b" = rep(x = binary, each = 50L)
)
lapply(object, class)
object <- encode(object)
lapply(object, class)
}
\seealso{
\code{Rle()}.
}
