% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,character,ANY,ANY-method}
\alias{export,matrix,ANY,ANY-method}
\alias{export,data.frame,ANY,ANY-method}
\alias{export,DataFrame,ANY,ANY-method}
\alias{export,Matrix,ANY,ANY-method}
\alias{export,GenomicRanges,ANY,ANY-method}
\title{Objects exported from other packages}
\usage{
export(object, con, format, ...)

\S4method{export}{character,ANY,ANY}(
  object,
  con,
  format,
  ext = c("txt", "txt.bz2", "txt.gz", "txt.xz", "txt.zip"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  append = FALSE,
  overwrite = NULL,
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{matrix,ANY,ANY}(
  object,
  con,
  format,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = NULL,
  engine = getOption(x = "acid.export.engine", "data.table"),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{data.frame,ANY,ANY}(
  object,
  con,
  format,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = NULL,
  engine = getOption(x = "acid.export.engine", "data.table"),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{DataFrame,ANY,ANY}(
  object,
  con,
  format,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = NULL,
  engine = getOption(x = "acid.export.engine", "data.table"),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{Matrix,ANY,ANY}(
  object,
  con,
  format,
  ext = c("mtx.gz", "mtx.bz2", "mtx.xz", "mtx.zip", "mtx"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  overwrite = NULL,
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{GenomicRanges,ANY,ANY}(
  object,
  con,
  format,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = NULL,
  engine = getOption(x = "acid.export.engine", "data.table"),
  quiet = getOption("acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.
An object supporting \code{dim()}, or a supported class capable of being coerced
to \code{data.frame}, to be written to disk.}

\item{con}{FIXME}

\item{format}{FIXME}

\item{ext}{\code{character(1)}.
Output file format extension.

\code{matrix} supported arguments:
\itemize{
\item Comma separated values (CSV):
\code{"csv"}, \code{"csv.bz2"}, \code{"csv.gz"}, \code{"csv.xz"}, \code{"csv.zip"}.
\item Tab separated values (TSV):
\code{"tsv"}, \code{"tsv.bz2"}, \code{"tsv.gz"}, \code{"tsv.xz"}, \code{"tsv.zip"}.
}

\code{Matrix} (\code{sparseMatrix}) supported arguments:
\itemize{
\item MatrixMarket exchange (MTX):
\code{"mtx"}, \code{"mtx.bz2"}, \code{"mtx.gz"}, \code{"mtx.xz"}, \code{"mtx.zip"}.
}}

\item{dir}{\code{character(1)}.
Directory path.}

\item{file}{\code{character(1)}.
File path. When left unset (default), the \code{ext} and \code{dir} arguments will
be used instead.}

\item{append}{\code{logical(1)}.
Append to output file.
When enabled, automatically sets \code{overwrite} argument to \code{FALSE}.
Requires readr package to be installed.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{engine}{\code{character(1)}.
Engine (package) to use for export.
Currently supported:
\itemize{
\item base
\item data.table
\item readr
\item vroom
}}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{IRanges}{\code{\link[IRanges:AtomicList-class]{CharacterList}}, \code{\link[IRanges:DataFrameList-class]{DataFrameList}}, \code{\link[IRanges:AtomicList-class]{FactorList}}, \code{\link[IRanges:IRanges-constructor]{IRanges}}, \code{\link[IRanges:AtomicList-class]{IntegerList}}, \code{\link[IRanges:AtomicList-class]{LogicalList}}, \code{\link[IRanges:AtomicList-class]{NumericList}}, \code{\link[IRanges:AtomicList-class]{RleList}}, \code{\link[IRanges:DataFrameList-class]{SplitDataFrameList}}}

  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}, \code{\link[S4Vectors:Factor-class]{Factor}}, \code{\link[S4Vectors:LLint-class]{LLint}}, \code{\link[S4Vectors:List-class]{List}}, \code{\link[S4Vectors:Rle-class]{Rle}}, \code{\link[S4Vectors:SimpleList-class]{SimpleList}}}
}
}
\note{
Updated 2021-09-24.
}
\section{Row names}{


Some export utilities in R have a tendency to drop row names when writing to
disk in CSV format. For example, the \href{https://readr.tidyverse.org/}{readr} family of functions never
write row names by design. This is a \emph{really poor} default setting for
handling genomic data, which often contain gene identifiers in the row names.
Here we're performing any internal tibble coercion step to ensure row names
are always moved to a \code{"rowname"} column in the CSV export.
}

\section{Debugging}{


Note that this function currently wraps \code{data.table::fwrite()} by default
for exporting \code{data.frame} and \code{matrix} class objects.
}

\examples{
counts <- matrix(data = seq_len(100L), nrow = 10)
export(counts, ext = "csv")

## Clean up.
file.remove("counts.csv")
}
\seealso{
Packages:
\itemize{
\item \href{http://r-datatable.com/}{data.table}.
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{https://cran.r-project.org/package=rio}{rio}.
\item \href{http://bioconductor.org/packages/rtracklayer/}{rtracklayer}.
\item \href{https://vroom.r-lib.org}{vroom}.
}

Export functions:
\itemize{
\item \code{data.table::fwrite()}.
\item \code{readr::write_csv()}.
\item \code{rio::export()}.
\item \code{rtracklayer::export()}.
\item \code{vroom::vroom_write()}.
}
}
