% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,character-method}
\alias{export,matrix-method}
\alias{export,data.frame-method}
\alias{export,DataFrame-method}
\alias{export,Matrix-method}
\alias{export,GenomicRanges-method}
\title{Export}
\usage{
export(object, ...)

\S4method{export}{character}(
  object,
  ext = "txt",
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  append = FALSE,
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{matrix}(
  object,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{data.frame}(
  object,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{DataFrame}(
  object,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{Matrix}(
  object,
  ext = "mtx",
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)

\S4method{export}{GenomicRanges}(
  object,
  ext = getOption("acid.export.ext", default = "csv"),
  dir = getOption("acid.export.dir", default = "."),
  file = NULL,
  rownames = TRUE,
  colnames = TRUE,
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.
An object supporting \code{\link[base:dim]{dim()}}, or a supported class capable
of being coerced to \code{data.frame}, to be written to disk.}

\item{ext}{\code{character(1)}.
Output file format extension.

\code{matrix} supported arguments:
\itemize{
\item Comma separated values (CSV):
\code{"csv"}, \code{"csv.bz2"}, \code{"csv.gz"}, \code{"csv.xz"}, \code{"csv.zip"}.
\item Tab separated values (TSV):
\code{"tsv"}, \code{"tsv.bz2"}, \code{"tsv.gz"}, \code{"tsv.xz"}, \code{"tsv.zip"}.
}

\code{Matrix} (\code{sparseMatrix}) supported arguments:
\itemize{
\item MatrixMarket exchange (MTX):
\code{"mtx"}, \code{"mtx.bz2"}, \code{"mtx.gz"}, \code{"mtx.xz"}, \code{"mtx.zip"}.
}}

\item{dir}{\code{character(1)}.
Directory path.}

\item{file}{\code{character(1)}.
File path. When left unset (default), the \code{ext} and \code{dir} arguments will
be used instead.}

\item{append}{\code{logical(1)}.
Append to output file.
When enabled, automatically sets \code{overwrite} argument to \code{FALSE}.
Requires readr package to be installed.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export data out of R and write to disk.
}
\note{
Updated 2021-03-16.
}
\section{Row names}{


Some export utilities in R have a tendency to drop row names when writing to
disk in CSV format. For example, the \href{https://readr.tidyverse.org/}{readr} family of functions never
write row names by design. This is a \emph{really poor} default setting for
handling genomic data, which often contain gene identifiers in the row names.
Here we're performing any internal tibble coercion step to ensure row names
are always moved to a \code{"rowname"} column in the CSV export.
}

\section{Debugging}{


Note that this function currently wraps \code{vroom::voom_write()} by default
for exporting \code{data.frame} and \code{matrix} class objects.
}

\examples{
counts <- matrix(data = seq_len(100L), nrow = 10)
export(counts, ext = "csv")

## Clean up.
file.remove("counts.csv")
}
\seealso{
Packages:
\itemize{
\item \href{http://r-datatable.com/}{data.table}.
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{https://cran.r-project.org/package=rio}{rio}.
\item \href{http://bioconductor.org/packages/rtracklayer/}{rtracklayer}.
\item \href{https://vroom.r-lib.org}{vroom}.
}

Export functions:
\itemize{
\item \code{data.table::fwrite()}.
\item \code{readr::write_csv()}.
\item \code{rio::export()}.
\item \code{rtracklayer::export()}.
\item \code{vroom::vroom_write()}.
}
}
