% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,ANY,missing-method}
\title{#' @rdname export
#' @export
setMethod(
f = "export",
signature = signature(
object = "DFrame",
con = "character",
format = "missing"
),
definition = \verb{export,DFrame}
)}
\usage{
export(object, con, ...)

\S4method{export}{ANY,missing}(object, con, ...)
}
\arguments{
\item{object}{Object.
An object supporting \code{dim()}, or a supported class capable of being coerced
to \code{data.frame}, to be written to disk.}

\item{con}{\code{character(1)}, \code{missing}, or \code{NULL}.
File path.
Alternatively, can leave unset and use \code{ext} and \code{dir} arguments instead.}

\item{...}{Additional arguments.}

\item{append}{\code{logical(1)}.
Append to output file.
When enabled, automatically sets \code{overwrite} argument to \code{FALSE}.
Requires readr package to be installed.}

\item{engine}{\code{character(1)}.
Engine (package) to use for export.

Currently supported:
\itemize{
\item base
\item data.table
\item readr
}}

\item{format}{\code{character(1)}, \code{missing}, or \code{NULL}.
\emph{Currently not supported.}}

\item{quote}{\code{logical(1)}.
Surround any \code{character} or \code{factor} columns by double quotes.
Recommended by default.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
#' @rdname export
#' @export
setMethod(
f = "export",
signature = signature(
object = "GRanges",
con = "character",
format = "missing"
),
definition = \verb{export,GRanges}
)
}
\details{
#' @rdname export
#' @export
setMethod(
f = "export",
signature = signature(
object = "GRangesList",
con = "character",
format = "missing"
),
definition = \verb{export,GRangesList}
)

#' @rdname export
#' @export
setMethod(
f = "export",
signature = signature(
object = "Matrix",
con = "character",
format = "missing"
),
definition = \verb{export,Matrix}
)

#' @rdname export
#' @export
setMethod(
f = "export",
signature = signature(
object = "atomic",
con = "character",
format = "missing"
),
definition = \verb{export,atomic}
)

#' @rdname export
#' @export
setMethod(
f = "export",
signature = signature(
object = "data.frame",
con = "character",
format = "missing"
),
definition = \verb{export,data.frame}
)

#' @rdname export
#' @export
setMethod(
f = "export",
signature = signature(
object = "matrix",
con = "character",
format = "missing"
),
definition = \verb{export,matrix}
)
}
\note{
Updated 2023-09-20.
}
\section{Output file format extension}{


\code{matrix} supported arguments:
\itemize{
\item Comma separated values (CSV):
\code{"csv"}, \code{"csv.bz2"}, \code{"csv.gz"}, \code{"csv.xz"}, \code{"csv.zip"}.
\item Tab separated values (TSV):
\code{"tsv"}, \code{"tsv.bz2"}, \code{"tsv.gz"}, \code{"tsv.xz"}, \code{"tsv.zip"}.
}

\code{Matrix} (\code{sparseMatrix}) supported arguments:
\itemize{
\item MatrixMarket exchange (MTX):
\code{"mtx"}, \code{"mtx.bz2"}, \code{"mtx.gz"}, \code{"mtx.xz"}, \code{"mtx.zip"}.
}
}

\section{Row names}{


Some export utilities in R have a tendency to drop row names when writing to
disk in CSV format. For example, the \href{https://readr.tidyverse.org/}{readr} family of functions never
write row names by design. This is a \emph{really poor} default setting for
handling genomic data, which often contain gene identifiers in the row names.
Here we're performing any internal tibble coercion step to ensure row names
are always moved to a \code{"rowname"} column in the CSV export.
}

\section{Debugging}{


Note that this function currently wraps \code{readr::write_delim()} by default
for exporting \code{DFrame}, \code{data.frame}, and \code{matrix} class objects.
}

\examples{
counts <- matrix(data = seq_len(100L), nrow = 10)
export(object = counts, con = "counts.csv")

## Clean up.
file.remove("counts.csv")
}
\seealso{
Packages:
\itemize{
\item \href{https://readr.tidyverse.org}{readr}.
\item \href{https://r-datatable.com/}{data.table}.
\item \href{https://cran.r-project.org/package=rio}{rio}.
\item \href{https://bioconductor.org/packages/rtracklayer/}{rtracklayer}.
\item \href{https://vroom.r-lib.org}{vroom}.
}

Export functions:
\itemize{
\item \code{BiocIO::export()}.
\item \code{data.table::fwrite()}.
\item \code{readr::write_delim()}.
\item \code{rio::export()}.
\item \code{rtracklayer::export()}.
\item \code{vroom::vroom_write()}.
}
}
