% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,DataFrame,character,missing-method}
\alias{export,GenomicRanges,character,missing-method}
\alias{export,Matrix,character,missing-method}
\alias{export,atomic,character,missing-method}
\alias{export,data.frame,character,missing-method}
\alias{export,matrix,character,missing-method}
\alias{export,ANY,missing,missing-method}
\alias{export,ANY,missing,character-method}
\title{Export}
\usage{
export(object, con, format, ...)

\S4method{export}{DataFrame,character,missing}(
  object,
  con,
  format,
  rownames = TRUE,
  colnames = TRUE,
  quote = TRUE,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{GenomicRanges,character,missing}(
  object,
  con,
  format,
  rownames = TRUE,
  colnames = TRUE,
  quote = TRUE,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{Matrix,character,missing}(
  object,
  con,
  format,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{atomic,character,missing}(
  object,
  con,
  format,
  append = FALSE,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption(x = "acid.quiet", default = FALSE),
  verbose = getOption(x = "acid.verbose", default = FALSE)
)

\S4method{export}{data.frame,character,missing}(
  object,
  con,
  format,
  rownames = TRUE,
  colnames = TRUE,
  quote = TRUE,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{matrix,character,missing}(
  object,
  con,
  format,
  rownames = TRUE,
  colnames = TRUE,
  quote = TRUE,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{ANY,missing,missing}(
  object,
  con,
  format,
  dir = getOption(x = "acid.export.dir", default = getwd()),
  ...
)
}
\arguments{
\item{object}{Object.
An object supporting \code{dim()}, or a supported class capable of being coerced
to \code{data.frame}, to be written to disk.}

\item{con}{\code{character(1)}, \code{missing}, or \code{NULL}.
File path.
Alternatively, can leave unset and use \code{ext} and \code{dir} arguments instead.}

\item{format}{\code{character(1)}, \code{missing}, or \code{NULL}.
\emph{Currently not supported.}}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{quote}{\code{logical(1)}.
Surround any \code{character} or \code{factor} columns by double quotes.
Recommended by default.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{engine}{\code{character(1)}.
Engine (package) to use for export.

Currently supported:
\itemize{
\item base
\item data.table
\item readr
}}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{append}{\code{logical(1)}.
Append to output file.
When enabled, automatically sets \code{overwrite} argument to \code{FALSE}.
Requires readr package to be installed.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export
}
\note{
Updated 2023-04-13.
}
\section{Output file format extension}{


\code{matrix} supported arguments:
\itemize{
\item Comma separated values (CSV):
\code{"csv"}, \code{"csv.bz2"}, \code{"csv.gz"}, \code{"csv.xz"}, \code{"csv.zip"}.
\item Tab separated values (TSV):
\code{"tsv"}, \code{"tsv.bz2"}, \code{"tsv.gz"}, \code{"tsv.xz"}, \code{"tsv.zip"}.
}

\code{Matrix} (\code{sparseMatrix}) supported arguments:
\itemize{
\item MatrixMarket exchange (MTX):
\code{"mtx"}, \code{"mtx.bz2"}, \code{"mtx.gz"}, \code{"mtx.xz"}, \code{"mtx.zip"}.
}
}

\section{Row names}{


Some export utilities in R have a tendency to drop row names when writing to
disk in CSV format. For example, the \href{https://readr.tidyverse.org/}{readr} family of functions never
write row names by design. This is a \emph{really poor} default setting for
handling genomic data, which often contain gene identifiers in the row names.
Here we're performing any internal tibble coercion step to ensure row names
are always moved to a \code{"rowname"} column in the CSV export.
}

\section{Debugging}{


Note that this function currently wraps \code{readr::write_delim()} by default
for exporting \code{DataFrame}, \code{data.frame}, and \code{matrix} class objects.
}

\examples{
counts <- matrix(data = seq_len(100L), nrow = 10)
export(object = counts, con = "counts.csv")

## Clean up.
file.remove("counts.csv")
}
\seealso{
Packages:
\itemize{
\item \href{https://readr.tidyverse.org}{readr}.
\item \href{https://r-datatable.com/}{data.table}.
\item \href{https://cran.r-project.org/package=rio}{rio}.
\item \href{https://bioconductor.org/packages/rtracklayer/}{rtracklayer}.
\item \href{https://vroom.r-lib.org}{vroom}.
}

Export functions:
\itemize{
\item \code{BiocIO::export()}.
\item \code{data.table::fwrite()}.
\item \code{readr::write_delim()}.
\item \code{rio::export()}.
\item \code{rtracklayer::export()}.
\item \code{vroom::vroom_write()}.
}
}
