% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/factorize-methods.R
\name{factorize}
\alias{factorize}
\alias{factorize,DFrame-method}
\alias{factorize,data.frame-method}
\title{Factorize}
\usage{
factorize(object, ...)

\S4method{factorize}{DFrame}(object)

\S4method{factorize}{data.frame}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
All columns will be coerced to \code{factor}.
}
\description{
Coerce all columns in the object to \code{factor} (i.e. \code{stringsAsFactors}).
}
\note{
Updated 2023-01-31.
}
\examples{
## DFrame ====
object <- S4Vectors::DataFrame(
    "a" = c("a", "b", "c", "d"),
    "b" = c("a", "a", "b", "b"),
    "c" = c(1L, 2L, 3L, 4L),
    "d" = c(1L, 2L, 1L, 2L),
    "e" = c(TRUE, TRUE, FALSE, FALSE),
    row.names = c("A", "B", "C", "D")
)
object <- factorize(object)
print(object)
}
\seealso{
\itemize{
\item Legacy \code{stringsAsFactors} approach for \code{data.frame} import.
}
}
