% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillLines.R
\name{fillLines}
\alias{fillLines}
\title{Fill delimited lines with NA values}
\usage{
fillLines(object, format = c("csv", "tsv"))
}
\arguments{
\item{object}{\code{character}.}

\item{format}{\code{character(1)}.
Delimited input format.}
}
\value{
\code{character}.
Modified character vector, with consistent number of delimited values.
}
\description{
Fill delimited lines with NA values
}
\note{
Updated 2023-09-27.
}
\examples{
## character ====
object <- c(
    "aaa\tbbb",
    "ccc\tddd\teee",
    "fff\tggg\thhh\tiii"
)
print(object)
object <- fillLines(object, format = "tsv")
print(object)
}
\seealso{
\itemize{
\item \code{textConnection}: For import approach after modification.
}
}
