% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJSON.R
\name{getJSON}
\alias{getJSON}
\title{Get JSON from an API}
\usage{
getJSON(url)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}
}
\value{
\code{list}.
}
\description{
Get JSON from an API
}
\note{
Updated 2023-07-06.
}
\examples{
## Access the UCSC Genome Browser API.
url <- "https://api.genome.ucsc.edu/list/ucscGenomes"
json <- getJSON(url)
names(json)
}
\seealso{
\itemize{
\item httr2 package.
}
}
