% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJson.R
\name{getJson}
\alias{getJson}
\title{Get JSON from an API}
\usage{
getJson(url)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}
}
\value{
\code{list}.
}
\description{
Get JSON from an API
}
\details{
Requires httr2 package to be installed.
}
\note{
Updated 2023-09-14.
}
\examples{
## Access the UCSC Genome Browser API.
url <- pasteUrl(
    "api.genome.ucsc.edu",
    "list",
    "ucscGenomes",
    protocol = "https"
)
if (goalie::isAnExistingUrl(url)) {
    json <- getJson(url)
    print(names(json))
}

## Access Ensembl genome assembly metadata.
## The Ensembl REST API server is prone to connection issues, so keeping
## disabled here in the working example.
## > url <- pasteUrl(
## >     "rest.ensembl.org",
## >     "info",
## >     "assembly",
## >     paste0(
## >         "Homo sapiens",
## >         "?", "content-type=application/json"
## >     ),
## >     protocol = "https"
## > )
## > if (goalie::isAnExistingUrl(url)) {
## >     json <- getJson(url)
## >     print(names(json))
## > }
}
