% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJSON.R
\name{getJSON}
\alias{getJSON}
\title{Get JSON from an API}
\usage{
getJSON(url)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}
}
\value{
\code{list}.
}
\description{
Get JSON from an API
}
\note{
Updated 2020-01-05.
}
\examples{
## Get current list of genomes from UCSC Genome Browser.
url <- "https://api.genome.ucsc.edu/list/ucscGenomes"
json <- getJSON(url)
names(json)
}
