% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURLDirList.R
\name{getURLDirList}
\alias{getURLDirList}
\title{Get remote URL directory listing}
\usage{
getURLDirList(url, pattern = NULL)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{pattern}{\code{character(1)}.
Regular expression pattern to use for matching.
Passes to \code{grepl} function internally.}
}
\value{
\code{character}.
Simple directory contents return, including both files and subdirectories.
}
\description{
Get remote URL directory listing
}
\details{
Requires RCurl package to be installed.
}
\note{
Best served using FTP instead of HTTPS.

Updated 2023-09-15.
}
\examples{
url <- "ftp://ftp.ncbi.nlm.nih.gov/genomes/"
if (goalie::isAnExistingURL(url)) {
    x <- getURLDirList(url)
    tail(x)
}
}
