% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURLDirList.R
\name{getURLDirList}
\alias{getURLDirList}
\title{Get remote URL directory listing}
\usage{
getURLDirList(
  url,
  type = c("all", "dirs", "files"),
  pattern = NULL,
  absolute = FALSE
)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{type}{\code{character(1)}.
Type of files to return:
\itemize{
\item \code{"all"}: Both directories and files.
\item \code{"dirs"}: Directories only.
\item \code{"files"}: Files only.
}}

\item{pattern}{\code{character(1)}.
Regular expression pattern to use for matching.
Passes to \code{grepl} function internally.}

\item{absolute}{\code{logical(1)}.
Return absolute path.}
}
\value{
\code{character}.
File basename, or absolute URL path when \code{absolute} is \code{TRUE}.
}
\description{
Get remote URL directory listing
}
\details{
Currently, only FTP servers are supported.
}
\note{
Updated 2023-09-18.
}
\examples{
url <- "ftp://ftp.ncbi.nlm.nih.gov/genomes/"
if (goalie::isAnExistingURL(url)) {
    x <- getURLDirList(url)
    print(x)
}
}
\seealso{
\itemize{
\item \code{curlGetHeaders}.
}
}
