% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURLDirList.R
\name{getURLDirList}
\alias{getURLDirList}
\title{Get remote URL directory listing}
\usage{
getURLDirList(url, pattern = NULL)
}
\arguments{
\item{url}{\code{character(1)}.
Uniform Resource Locator (URL).}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}
}
\value{
\code{character}.
Simple directory contents return, including both files and subdirectories.
}
\description{
Get remote URL directory listing
}
\note{
Best served using FTP instead of HTTP.

Updated 2019-10-18.
}
\examples{
if (
    hasInternet() &&
    !isTRUE(nzchar(Sys.getenv("CI")))
) {
    url <- "ftp://ftp.ensembl.org/pub/"
    x <- getURLDirList(url, pattern = "^release-")
    tail(x)
}
}
