% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load data}
\usage{
loadData(
  ...,
  dir = getOption("acid.load.dir", default = "."),
  envir = globalenv(),
  list = NULL,
  overwrite = getOption("acid.overwrite", default = TRUE)
)
}
\arguments{
\item{...}{Object names.
Note that these arguments are interpreted as symbols using non-standard
evaluation for convenience during interactive use, and \emph{must not be
quoted}.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{envir}{\code{environment}.
Environment.}

\item{list}{\code{character}.
A character vector containing the names of objects to be loaded.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}
}
\value{
Invisible \code{character}.
File paths.
}
\description{
Load R data files from a directory using symbols rather than complete file
paths. Supports \code{RDS}, \code{RDA}, and \code{RDATA} file extensions.
}
\details{
\code{loadData()} is opinionated about the format of R data files it will accept.
\code{save()} allows for the saving of multiple objects into a single R data file.
This can later result in unexpected accidental replacement of an existing
object in the current environment. Since an R data file internally stores the
name of an object, if the file is later renamed the object name will no
longer match.

To avoid any accidental replacements, \code{loadData()} will only load R data
files that contain a single object, and the internal object name must match
the file name exactly. Additionally, \code{loadData()} will intentionally error if
an object with the same name already exists in the destination \code{environment}.
}
\note{
This function is desired for interactive use and interprets object
names using non-standard evaluation.

Updated 2021-08-24.
}
\examples{
dir <- system.file("extdata", package = "pipette")

## Interactive mode ====
## Note that this method uses non-standard evaluation.
loadData(example, dir = dir)

## Clean up.
rm(example, inherits = TRUE)

## List mode ====
## Note that this method uses standard evaluation.
## Use this approach inside of functions.
list <- "example"
loadData(list = list, dir = dir)

## Clean up.
rm(example, inherits = TRUE)
}
\seealso{
\itemize{
\item \code{load()}.
\item \code{readRDS()}.
}
}
