% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically handle a local or remote file path}
\usage{
localOrRemoteFile(file)
}
\arguments{
\item{file}{\code{character(1)}.
Local file paths or remote URLs.}
}
\value{
\code{character}.
Local file path(s). Stops on a missing file.
}
\description{
Dynamically handle a local or remote file path
}
\section{Vectorization}{


This function is vectorized and supports mixed local and remote paths. Remote
files are downloaded locally to a temporary directory.
}

\section{Compressed files}{


Compressed files will automatically be decompressed. Currently, these file
extensions are natively supported: \code{BZ2}, \code{GZ}, \code{XZ}, \code{ZIP}.
}

\examples{
## Local
file <- system.file("extdata/example.csv", package = "brio")
x <- localOrRemoteFile(file)
basename(x)

## Remote
file <- "http://basejump.seq.cloud/v0.9/hgnc.txt.gz"
x <- localOrRemoteFile(file)
basename(x)
}
\seealso{
\code{\link[=tempdir]{tempdir()}}.
}
