% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically handle a local or remote file path}
\usage{
localOrRemoteFile(file)
}
\arguments{
\item{file}{\code{character(1)}.
Local file paths or remote URLs.}
}
\value{
\code{character}.
Local file path(s). Stops on a missing file.
}
\description{
Dynamically handle a local or remote file path
}
\note{
This function is vectorized and supports mixed local and remote paths. Remote
files are downloaded locally to a temporary directory.
}
\examples{
file <- system.file("extdata/example.csv", package = "brio")
x <- localOrRemoteFile(file)
basename(x)
}
\seealso{
\code{\link[=tempdir]{tempdir()}}.
}
