% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOrRemoteFile.R
\name{localOrRemoteFile}
\alias{localOrRemoteFile}
\title{Dynamically handle a local or remote file path}
\usage{
localOrRemoteFile(file, quiet = getOption("acid.quiet", default = FALSE))
}
\arguments{
\item{file}{\code{character(1)}.
Local file paths or remote URLs.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}
}
\value{
\code{character}.
Local file path(s). Stops on a missing file.
}
\description{
Dynamically handle a local or remote file path
}
\note{
Updated 2020-12-10.
}
\section{Vectorization}{


This function is vectorized and supports mixed local and remote paths. Remote
files are downloaded locally to a temporary directory.
}

\section{Compressed files}{


Compressed files will automatically be decompressed. Currently, these file
extensions are natively supported: \code{BZ2}, \code{GZ}, \code{XZ}, and \code{ZIP}.
}

\examples{
## Local
file <- system.file("extdata/example.csv", package = "pipette")
x <- localOrRemoteFile(file)
basename(x)

## Remote
file <- AcidBase::pasteURL(
    pipetteTestsURL,
    "hgnc.txt.gz",
    protocol = "none"
)
x <- localOrRemoteFile(file)
basename(x)
}
\seealso{
\itemize{
\item \code{\link[=tempfile]{tempfile()}}.
\item \code{\link[=tempdir]{tempdir()}}.
}
}
