% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{naStrings}
\alias{naStrings}
\title{NA strings}
\format{
An object of class \code{character} of length 12.
}
\usage{
naStrings
}
\description{
NA strings
}
\note{
Updated 2023-04-12.
}
\examples{
naStrings
}
\keyword{datasets}
