% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PipetteFile-class}
\alias{PipetteFile-class}
\alias{PipetteDelimFile-class}
\alias{PipetteExcelFile-class}
\alias{PipetteLinesFile-class}
\alias{PipetteRioFile-class}
\alias{PipetteRtracklayerFile-class}
\alias{PipetteBAMFile-class}
\alias{PipetteBCFFile-class}
\alias{PipetteBcbioCountsFile-class}
\alias{PipetteCRAMFile-class}
\alias{PipetteCSVFile-class}
\alias{PipetteFASTAFile-class}
\alias{PipetteFASTQFile-class}
\alias{PipetteGCTFile-class}
\alias{PipetteGMTFile-class}
\alias{PipetteGMXFile-class}
\alias{PipetteGRPFile-class}
\alias{PipetteJSONFile-class}
\alias{PipetteMAFFile-class}
\alias{PipetteMTXFile-class}
\alias{PipetteOBOFile-class}
\alias{PipettePZFXFile-class}
\alias{PipetteRDataFile-class}
\alias{PipetteSAMFile-class}
\alias{PipetteRDSFile-class}
\alias{PipetteTableFile-class}
\alias{PipetteTSVFile-class}
\alias{PipetteVCFFile-class}
\alias{PipetteYAMLFile-class}
\title{File extension classes}
\description{
Currently intended for use with \code{import} function.
}
\details{
Extends \code{BiocFile} defined in BiocIO package.
}
\note{
Updated 2023-07-12.
}
\section{Primary classes}{

\itemize{
\item \code{PipetteBAMFile}: Binary sequence alignment map (BAM).
\item \code{PipetteBCFFile}: Binary variant call format (BCF).
\item \code{PipetteBcbioCountsFile}: bcbio-nextgen counts file.
\item \code{PipetteCRAMFile}: Compressed reference-oriented alignment map (CRAM).
\item \code{PipetteCSVFile}: Comma-separated values file (CSV).
\item \code{PipetteFASTAFile}: FASTA file.
\item \code{PipetteFASTQFile}: FASTQ file.
\item \code{PipetteGCTFile}: Gene Cluster Text file (GCT).
\item \code{PipetteGMTFile}: Gene matrix transposed file (GMT).
\item \code{PipetteGMXFile}: Gene matrix file (GMX).
\item \code{PipetteGRPFile}: Gene set file (GRP).
\item \code{PipetteJSONFile}: JSON file.
\item \code{PipetteMAFFile}: Mutation annotation format file (MAF).
\item \code{PipetteMTXFile}: MatrixMarket exchange file (MTX).
\item \code{PipetteOBOFile}: Open Biomedical Ontologies file (OBO).
\item \code{PipettePZFXFile}: GraphPad Prism file (PZFX).
\item \code{PipetteRDSFile}: R data file containing a single, serialized object (RDS).
\item \code{PipetteRDataFile}: R Data file containing multiple objects (RData/RDA).
\item \code{PipetteSAMFile}: Sequence alignment map (SAM).
\item \code{PipetteTSVFile}: Tab-separated values file (TSV).
\item \code{PipetteTableFile}: Base R table file (TXT).
\item \code{PipetteVCFFile}: Variant call format (VCF).
\item \code{PipetteYAMLFile}: YAML file.
}
}

\section{\code{PipetteDelimFile-class}}{


Delimited file.

File extension group supporting:
\itemize{
\item CSV: Comma-separated values
\item TSV: Tab-separated values
\item Base R table (TXT)
}
}

\section{\code{PipetteExcelFile-class}}{


Microsoft Excel file.

File extension group supporting:
\itemize{
\item XLS
\item XLSB
\item XLSX
}
}

\section{\code{PipetteLinesFile-class}}{


Source code lines file.

File extension group supporting:
\itemize{
\item BASH: Bash shell script
\item LOG: Log file
\item MD: Markdown file
\item PY: Python script
\item R: R script
\item RMD: R Markdown file
\item SH: Shell script (e.g. Bash, POSIX)
\item ZSH: Zsh shell script
}
}

\section{\code{PipetteRioFile-class}}{


File extension supported by \code{rio::import}.

File extension group supporting:
\itemize{
\item ARFF: Weka Attribute-Relation File Format
\item DBF: dBase Database File
\item DIF: Data Interchange Format
\item DTA: Stata
\item FWF: Fixed-width formatting
\item MAT: Matlab
\item MTP: Minitab
\item ODS: OpenDocument (LibreOffice)
\item POR: SPSS
\item PSV: Pipe-separated values
\item SAS7BDAT: SASS
\item SAV: SPSS
\item SYD: Systat
\item REC: Epi Info
\item XPT: SASS
}
}

\section{\code{PipetteRtracklayerFile-class}}{


File extension supported by \code{rtracklayer::import}.

File extension group supporting:
\itemize{
\item BED, BED15, BEDGRAPH, BEDPE
\item BIGWIG, BW, WIG
\item GFF, GFF1, GFF2, GFF3, GTF
\item BROADPEAK, NARROWPEAK
}
}

\seealso{
\itemize{
\item \code{getGeneric}.
\item \code{BiocIO::FileForFormat}.
}
}
