% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PipetteFile-class}
\alias{PipetteFile-class}
\alias{DelimFile-class}
\alias{ExcelFile-class}
\alias{LinesFile-class}
\alias{RioFile-class}
\alias{RtracklayerFile-class}
\alias{BcbioCountsFile-class}
\alias{CSVFile-class}
\alias{FASTAFile-class}
\alias{FASTQFile-class}
\alias{GMTFile-class}
\alias{GMXFile-class}
\alias{GRPFile-class}
\alias{JSONFile-class}
\alias{MTXFile-class}
\alias{PZFXFile-class}
\alias{RDataFile-class}
\alias{RDSFile-class}
\alias{TableFile-class}
\alias{TSVFile-class}
\alias{YAMLFile-class}
\title{File extension classes}
\description{
Currently intended for use with \code{import} function.
}
\section{Functions}{
\itemize{
\item \code{DelimFile-class}: Delimited file.\cr
File extension group supporting:
\itemize{
\item CSV: Comma-separated values
\item TSV: Tab-separated values
\item Base R table (TXT)
}

\item \code{ExcelFile-class}: Microsoft Excel file.\cr
File extension group supporting:
\itemize{
\item XLS
\item XLSB
\item XLSX
}

\item \code{LinesFile-class}: Source code lines file.\cr
File extension group supporting:
\itemize{
\item BASH: Bash shell script
\item LOG: Log file
\item MD: Markdown file
\item PY: Python script
\item R: R script
\item RMD: R Markdown file
\item SH: Shell script (e.g. Bash, POSIX)
\item ZSH: Zsh shell script
}

\item \code{RioFile-class}: File extension supported by \code{rio::import}.\cr
File extension group supporting:
\itemize{
\item ARFF: Weka Attribute-Relation File Format
\item DBF: dBase Database File
\item DIF: Data Interchange Format
\item DTA: Stata
\item FWF: Fixed-width formatting
\item MAT: Matlab
\item MTP: Minitab
\item ODS: OpenDocument (LibreOffice)
\item POR: SPSS
\item PSV: Pipe-separated values
\item SAS7BDAT: SASS
\item SAV: SPSS
\item SYD: Systat
\item REC: Epi Info
\item XPT: SASS
}

\item \code{RtracklayerFile-class}: File extension supported by \code{rtracklayer::import}.\cr
File extension group supporting:
\itemize{
\item BED, BED15, BEDGRAPH, BEDPE
\item BIGWIG, BW, WIG
\item GFF, GFF1, GFF2, GFF3, GTF
\item BROADPEAK, NARROWPEAK
}

\item \code{BcbioCountsFile-class}: bcbio-nextgen counts file.

\item \code{CSVFile-class}: Comma-separated values file (CSV).

\item \code{FASTAFile-class}: FASTA file.

\item \code{FASTQFile-class}: FASTQ file.

\item \code{GMTFile-class}: Gene matrix transposed file (GMT).

\item \code{GMXFile-class}: Gene matrix file (GMX).

\item \code{GRPFile-class}: Gene set file (GRP).

\item \code{JSONFile-class}: JSON file.

\item \code{MTXFile-class}: MatrixMarket exchange file (MTX).

\item \code{PZFXFile-class}: GraphPad Prism file (PZFX).

\item \code{RDataFile-class}: R Data file containing multiple objects (RData/RDA).

\item \code{RDSFile-class}: R data file containing a single, serialized object (RDS).

\item \code{TableFile-class}: Base R table file (TXT).

\item \code{TSVFile-class}: Tab-separated values file (TSV).

\item \code{YAMLFile-class}: YAML file.
}}

\note{
Updated 2021-09-24.
}
\seealso{
\itemize{
\item \code{getGeneric}.
}
}
\keyword{internal}
