% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{pipetteTestsURL}
\alias{pipetteTestsURL}
\title{pipette test data URL}
\format{An object of class \code{character} of length 1.}
\usage{
pipetteTestsURL
}
\description{
pipette test data URL
}
\examples{
pipetteTestsURL
}
\keyword{internal}
