% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pasteURL}
\alias{getURL}
\alias{makeNames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{pasteURL}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{syntactic}{\code{\link[syntactic]{makeNames}}}
}}

