% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R, R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compressFile}
\alias{decompressFile}
\alias{bzip2}
\alias{bunzip2}
\alias{gzip}
\alias{gunzip}
\alias{tar}
\alias{untar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{export}}}

  \item{R.utils}{\code{\link[R.utils]{compressFile}}, \code{\link[R.utils]{decompressFile}}, \code{\link[R.utils]{bzip2}}, \code{\link[R.utils]{bunzip2}}, \code{\link[R.utils]{gzip}}, \code{\link[R.utils]{gunzip}}}

  \item{utils}{\code{\link[utils]{tar}}, \code{\link[utils]{untar}}}
}}

