% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DataFrame}
\alias{GRanges}
\alias{GRangesList}
\alias{Matrix}
\alias{column_to_rownames}
\alias{compress}
\alias{data.table}
\alias{decompress}
\alias{download}
\alias{getURL}
\alias{makeNames}
\alias{metadata}
\alias{metadata<-}
\alias{pasteURL}
\alias{rbindlist}
\alias{rownames_to_column}
\alias{sparseMatrix}
\alias{tar}
\alias{tibble}
\alias{untar}
\alias{Matrix-class}
\alias{sparseMatrix-class}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{compress}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase]{pasteURL}}, \code{\link[AcidBase:reexports]{tar}}, \code{\link[AcidBase:reexports]{untar}}}

  \item{AcidGenerics}{\code{\link[AcidGenerics:reexports]{DataFrame}}, \code{\link[AcidGenerics:reexports]{GRanges}}, \code{\link[AcidGenerics:reexports]{GRangesList}}, \code{\link[AcidGenerics:reexports]{metadata}}, \code{\link[AcidGenerics:reexports]{metadata<-}}}

  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{rbindlist}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}, \code{\link[Matrix]{sparseMatrix}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{syntactic}{\code{\link[syntactic]{makeNames}}}

  \item{tibble}{\code{\link[tibble:rownames]{column_to_rownames}}, \code{\link[tibble:rownames]{rownames_to_column}}, \code{\link[tibble]{tibble}}}
}}

