% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R, R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gzip}
\alias{untar}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{export}}}

  \item{R.utils}{\code{\link[R.utils]{gzip}}}

  \item{utils}{\code{\link[utils]{untar}}}
}}

