% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{CharacterList}
\alias{DataFrame}
\alias{DataFrameList}
\alias{FactorList}
\alias{GRanges}
\alias{GRangesList}
\alias{IRanges}
\alias{IntegerList}
\alias{LogicalList}
\alias{Matrix}
\alias{NumericList}
\alias{RleList}
\alias{SplitDataFrameList}
\alias{column_to_rownames}
\alias{compress}
\alias{data.table}
\alias{decompress}
\alias{download}
\alias{getURL}
\alias{makeNames}
\alias{metadata}
\alias{metadata<-}
\alias{pasteURL}
\alias{ranges}
\alias{rbindlist}
\alias{rownames_to_column}
\alias{sparseMatrix}
\alias{tibble}
\alias{trim}
\alias{AtomicList-class}
\alias{CharacterList-class}
\alias{DataFrameList-class}
\alias{FactorList-class}
\alias{GRanges-class}
\alias{GRangesList-class}
\alias{IRanges-class}
\alias{IntegerList-class}
\alias{LogicalList-class}
\alias{Matrix-class}
\alias{NumericList-class}
\alias{RleList-class}
\alias{SplitDataFrameList-class}
\alias{sparseMatrix-class}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{compress}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase]{pasteURL}}}

  \item{AcidGenerics}{\code{\link[AcidGenerics:reexports]{DataFrame}}, \code{\link[AcidGenerics:reexports]{metadata}}, \code{\link[AcidGenerics:reexports]{metadata<-}}}

  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{rbindlist}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges:GRanges-class]{GRanges}}, \code{\link[GenomicRanges:GRangesList-class]{GRangesList}}}

  \item{IRanges}{\code{\link[IRanges:AtomicList-class]{CharacterList}}, \code{\link[IRanges:DataFrameList-class]{DataFrameList}}, \code{\link[IRanges:AtomicList-class]{FactorList}}, \code{\link[IRanges:IRanges-constructor]{IRanges}}, \code{\link[IRanges:AtomicList-class]{IntegerList}}, \code{\link[IRanges:AtomicList-class]{LogicalList}}, \code{\link[IRanges:AtomicList-class]{NumericList}}, \code{\link[IRanges:AtomicList-class]{RleList}}, \code{\link[IRanges:DataFrameList-class]{SplitDataFrameList}}, \code{\link[IRanges:range-squeezers]{ranges}}, \code{\link[IRanges:Views-class]{trim}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}, \code{\link[Matrix]{sparseMatrix}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{syntactic}{\code{\link[syntactic]{makeNames}}}

  \item{tibble}{\code{\link[tibble:rownames]{column_to_rownames}}, \code{\link[tibble:rownames]{rownames_to_column}}, \code{\link[tibble]{tibble}}}
}}

