% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{coerce}
\alias{compress}
\alias{decompress}
\alias{download}
\alias{pasteURL}
\alias{getURL}
\alias{DataFrame}
\alias{data.table}
\alias{rbindlist}
\alias{makeNames}
\alias{column_to_rownames}
\alias{rownames_to_column}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{compress}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase]{pasteURL}}}

  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{rbindlist}}}

  \item{methods}{\code{\link[methods:setAs]{coerce}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}}

  \item{syntactic}{\code{\link[syntactic]{makeNames}}}

  \item{tibble}{\code{\link[tibble:rownames]{column_to_rownames}}, \code{\link[tibble:rownames]{rownames_to_column}}, \code{\link[tibble]{tibble}}}
}}

