% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize-methods.R, R/decode-methods.R,
%   R/encode-methods.R, R/export-methods.R, R/factorize-methods.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bzip2}
\alias{bunzip2}
\alias{gzip}
\alias{gunzip}
\alias{tar}
\alias{untar}
\alias{zip}
\alias{unzip}
\alias{isCompressedFile}
\alias{isBzipped}
\alias{isGzipped}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{atomize}}, \code{\link[bioverbs]{encode}}, \code{\link[bioverbs]{export}}, \code{\link[bioverbs]{factorize}}}

  \item{R.utils}{\code{\link[R.utils]{bzip2}}, \code{\link[R.utils]{bunzip2}}, \code{\link[R.utils]{gzip}}, \code{\link[R.utils]{gunzip}}, \code{\link[R.utils]{isCompressedFile}}, \code{\link[R.utils]{isBzipped}}, \code{\link[R.utils]{isGzipped}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{decode}}}

  \item{utils}{\code{\link[utils]{tar}}, \code{\link[utils]{untar}}, \code{\link[utils]{zip}}, \code{\link[utils]{unzip}}}
}}

