% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/removeNA-methods.R,
%   R/sanitizeNA-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compressFile}
\alias{decompressFile}
\alias{bzip2}
\alias{bunzip2}
\alias{gzip}
\alias{gunzip}
\alias{tar}
\alias{untar}
\alias{zip}
\alias{unzip}
\alias{isCompressedFile}
\alias{isBzipped}
\alias{isGzipped}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{removeNA}}, \code{\link[bioverbs]{sanitizeNA}}}

  \item{R.utils}{\code{\link[R.utils]{compressFile}}, \code{\link[R.utils]{decompressFile}}, \code{\link[R.utils]{bzip2}}, \code{\link[R.utils]{bunzip2}}, \code{\link[R.utils]{gzip}}, \code{\link[R.utils]{gunzip}}, \code{\link[R.utils]{isCompressedFile}}, \code{\link[R.utils]{isBzipped}}, \code{\link[R.utils]{isGzipped}}}

  \item{utils}{\code{\link[utils]{tar}}, \code{\link[utils]{untar}}, \code{\link[utils]{zip}}, \code{\link[utils]{unzip}}}
}}

