% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{compress}
\alias{decompress}
\alias{download}
\alias{pasteURL}
\alias{GRanges}
\alias{GRangesList}
\alias{DataFrameList}
\alias{IRanges}
\alias{SplitDataFrameList}
\alias{Matrix}
\alias{sparseMatrix}
\alias{getURL}
\alias{DataFrame}
\alias{List}
\alias{Rle}
\alias{SimpleList}
\alias{data.table}
\alias{rbindlist}
\alias{makeNames}
\alias{column_to_rownames}
\alias{rownames_to_column}
\alias{tibble}
\alias{DataFrame-class}
\alias{DataFrameList-class}
\alias{GRanges-class}
\alias{GRangesList-class}
\alias{IRanges-class}
\alias{List-class}
\alias{Matrix-class}
\alias{Rle-class}
\alias{SimpleList-class}
\alias{SplitDataFrameList-class}
\alias{sparseMatrix-class}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{compress}}, \code{\link[AcidBase:compress]{decompress}}, \code{\link[AcidBase]{download}}, \code{\link[AcidBase]{pasteURL}}}

  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{rbindlist}}}

  \item{GenomicRanges}{\code{\link[GenomicRanges:GRanges-class]{GRanges}}, \code{\link[GenomicRanges:GRangesList-class]{GRangesList}}}

  \item{IRanges}{\code{\link[IRanges:DataFrameList-class]{DataFrameList}}, \code{\link[IRanges:IRanges-constructor]{IRanges}}, \code{\link[IRanges:DataFrameList-class]{SplitDataFrameList}}}

  \item{Matrix}{\code{\link[Matrix]{Matrix}}, \code{\link[Matrix]{sparseMatrix}}}

  \item{RCurl}{\code{\link[RCurl]{getURL}}}

  \item{S4Vectors}{\code{\link[S4Vectors:DataFrame-class]{DataFrame}}, \code{\link[S4Vectors:List-class]{List}}, \code{\link[S4Vectors:Rle-class]{Rle}}, \code{\link[S4Vectors:SimpleList-class]{SimpleList}}}

  \item{syntactic}{\code{\link[syntactic]{makeNames}}}

  \item{tibble}{\code{\link[tibble:rownames]{column_to_rownames}}, \code{\link[tibble:rownames]{rownames_to_column}}, \code{\link[tibble]{tibble}}}
}}

