% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{transmit}
\alias{transmit}
\title{Transmit files from a remote FTP server}
\usage{
transmit(
  remoteDir,
  localDir = ".",
  pattern,
  rename = NULL,
  compress = FALSE,
  download = TRUE
)
}
\arguments{
\item{remoteDir}{\code{character(1)}.
Remote FTP directory path.}

\item{localDir}{\code{character(1)}.
Directory where to save files locally.}

\item{pattern}{\code{character(1)}.
Pattern to use for matching.}

\item{rename}{\code{character(1)} or \code{NULL}.
Rename the local files (including suffix), if desired.}

\item{compress}{\code{logical(1)}.
Compress the downloaded files.}

\item{download}{\code{logical(1)}.
Download files (default) or only return the matching URL(s).}
}
\value{
Invisible \code{character}.
Local file paths.
}
\description{
Utility function that supports easy file matching and download from a remote
FTP server. Also enables on-the-fly file renaming and compression.
}
\note{
Updated 2021-02-02.
}
\examples{
try({
    remoteDir <- "ftp://ftp.ncbi.nlm.nih.gov/genomes/"
    readme <- transmit(
        remoteDir = remoteDir,
        pattern = "^README\\\\.txt$",
        rename = "ncbi-readme.txt",
        compress = FALSE
    )
    basename(readme)
    file.exists(readme)
    unlink(readme)
})
}
